Added review
ReadCharger -> ReadChargers
locationId is private
added setAddress method
added setId method
duplicates status is a bool now
exception is being thrown when checking for duplicate locations
Stepdefs check the exceptionmessage
Added update location address test and stepdefs
Added delete a non existent location test and stepdefs

+++ ElectricChargingStation/src/main/java/org/example/LocationService.java
@@ -6,32 +6,32 @@ import java.util.Map;
 public class LocationService {
     public Map<Integer, Location> locationRepo = new HashMap<>();
     public String allLocations = "";
-    public String dupesStatus = "No dupes";
+    public boolean duplicates = false;
 
-    public void createLocation(String name, String address) {
+    public void createLocation(String name, String address) throws Exception {
         checkDupes(name, address);
-        if(dupesStatus.equals("No dupes")){
+        if(!duplicates){
             int newId = locationRepo.size() + 1;
             locationRepo.put(newId, new Location(name, address));
-            locationRepo.get(newId).locationId = newId;
+            locationRepo.get(newId).setId(newId);
         }
 
     }
 
-    public void checkDupes(String name, String address){
+    public void checkDupes(String name, String address) throws Exception {
         for (Map.Entry<Integer, Location> entry: locationRepo.entrySet()){
             if(entry.getValue().getName().equals(name)){
-                dupesStatus = "Location already exists";
-                return;
+                duplicates = true;
+                throw new Exception("Location already exists");
             }
             if(entry.getValue().getAddress().equals(address)){
-                dupesStatus = "Location already exists";
-                return;
+                duplicates = true;
+                throw new Exception("Location already exists");
             }
         }
     }
 
-    public String getAllLocations(){
+    public String getAllLocationsAsString(){
         StringBuilder actualOutput = new StringBuilder();
         for (Location location : locationRepo.values()) {
             actualOutput.append(location.toString()).append("\n");
@@ -39,11 +39,19 @@ public class LocationService {
         return actualOutput.toString();
     }
 
-    public void updateLocation(int id, String name){
+    public void updateLocationName(int id, String name){
         locationRepo.get(id).setName(name);
     }
 
-    public void deleteLocataion(int id){
-        locationRepo.remove(id);
+    public void updateLocationAddress(int id, String address){
+        locationRepo.get(id).setAddress(address);
+    }
+
+    public void deleteLocataion(int id) throws Exception {
+        if (locationRepo.containsKey(id)){
+            locationRepo.remove(id);
+        } else{
+            throw new Exception("Location not found");
+        }
     }
 }
